// src/models/enums/phases.ts
export enum GamePhase {
  SETUP = 'SETUP',
  CHARACTER_SELECTION = 'CHARACTER_SELECTION',
  SEASONAL_ASSESSMENT = 'SEASONAL_ASSESSMENT',
  THREAT_LEVEL_CHECK = 'THREAT_LEVEL_CHECK',
  LANDSCAPE_CHALLENGE = 'LANDSCAPE_CHALLENGE',
  CHALLENGE_RESOLUTION = 'CHALLENGE_RESOLUTION',
  RESOURCE_MANAGEMENT = 'RESOURCE_MANAGEMENT',
  ANIMAL_COMPANION = 'ANIMAL_COMPANION',
  CRAFTING = 'CRAFTING',
  JOURNEY_PROGRESSION = 'JOURNEY_PROGRESSION',
  HEALING_RECOVERY = 'HEALING_RECOVERY',
  GAME_OVER = 'GAME_OVER'
}